Rename vm module to evm
