+++ lib/evm/memory.js
