+++ lib/evm/opFns.js
