+++ lib/evm/opcodes.js
