+++ lib/evm/stack.js
