+++ lib/runCode.js
@@ -15,10 +15,10 @@ const Buffer = require('safe-buffer').Buffer
 const async = require('async')
 const utils = require('ethereumjs-util')
 const Block = require('ethereumjs-block')
-const lookupOpInfo = require('./vm/opcodes.js')
-const opFns = require('./vm/opFns.js')
-const Memory = require('./vm/memory')
-const Stack = require('./vm/stack')
+const lookupOpInfo = require('./evm/opcodes.js')
+const opFns = require('./evm/opFns.js')
+const Memory = require('./evm/memory')
+const Stack = require('./evm/stack')
 const exceptions = require('./exceptions.js')
 const { StorageReader } = require('./state')
 const setImmediate = require('timers').setImmediate
